tests/requirements.txt
pytest==7.2.2
pytest-asyncio==0.21.0
python-dotenv==0.21.1
redis==4.5.1
aiohttp==3.8.4
pydantic==1.10.5
requests==2.28.2